Create README.md
To be honest, whether lost bitcoin can be found or not depends on how it was lost and because of this, one needs a professional recovery expert like Fastfund Rcovery who will help you get back your lost bitcoin/funds. This Professional recovery Agent helped me recover my stolen 95,440 USD worth of bitcoin effortlessly. His service is not only quick but professional and reliable. I contacted Fastfund Recovery when I had invested in a fake crypto investment platform that cost me 95,440 USD Two months back. I saw an ad on Google on how one can invest on a certain investment platform and earn more and I thought the investment project was a great one that would yield me massive profit but I was wrong. I did invest because I had some cash saved for such an opportunity. I checked my account and I was glad that I indeed made a profit and now the problem came when it was time to withdraw the cash. I tried to withdraw the money for 3 days and that was when I discovered that I had to add more cash for the withdrawal. I promised myself I wouldn’t do this, and that’s when I contacted Fastfund Recovery. I discovered him through a positive review I saw and reached out to him via email. It was such a huge relief when he was able to recover my bitcoin funds. Contact his team today on: Fastfundrecovery8(AT)gmail(DOT)com and W/H: (+1) (903) (717) (6241)